refactor: simplify and DRY up extracted components
- Create unified mediaUtils for all media extraction logic
- Merge InlineNprofile and InlineNostrReference into single InlineNostrToken component
- Simplify NoteMedia component using shared utilities
- Remove redundant props and helper functions
- Delete duplicate components to keep code DRY

Reduces code duplication and improves maintainability.
